mistery

+++ include/Cell.h
@@ -1,8 +1,9 @@
 #pragma once
+#include <vector>
 
 struct Cell {
   std::vector<int> points_idx;
-  int8_t status; // predictable / unpredictable
-  int8_t label;  // ground truth: traversable / not traversable
+  int status; // predictable / unpredictable
+  int label;  // ground truth: traversable / not traversable
   float predicted_label; // prediction: traversable / not traversable
 };
